+++ service/conversations/v1/configuration/api_op_client.go
@@ -3,6 +3,8 @@ package configuration
 
 import (
 	"github.com/RJPearson94/twilio-sdk-go/client"
+	"github.com/RJPearson94/twilio-sdk-go/service/conversations/v1/configuration/address"
+	"github.com/RJPearson94/twilio-sdk-go/service/conversations/v1/configuration/addresses"
 	"github.com/RJPearson94/twilio-sdk-go/service/conversations/v1/configuration/webhook"
 )
 
@@ -10,7 +12,9 @@ import (
 type Client struct {
 	client *client.Client
 
-	Webhook func() *webhook.Client
+	Address   func(string) *address.Client
+	Addresses *addresses.Client
+	Webhook   func() *webhook.Client
 }
 
 // New creates a new instance of the configuration client
@@ -18,6 +22,12 @@ func New(client *client.Client) *Client {
 	return &Client{
 		client: client,
 
-		Webhook: func() *webhook.Client { return webhook.New(client) },
+		Address: func(addressSid string) *address.Client {
+			return address.New(client, address.ClientProperties{
+				Sid: addressSid,
+			})
+		},
+		Addresses: addresses.New(client),
+		Webhook:   func() *webhook.Client { return webhook.New(client) },
 	}
 }